Move components into components folder
